+++ Frameworks/libcelt_0061/celt-0.6.1/libcelt/kiss_fftr.c
@@ -45,7 +45,7 @@ kiss_fftr_cfg kiss_fftr_alloc(int nfft,void * mem,size_t * lenmem)
     memneeded = sizeof(struct kiss_fftr_state) + subsize + sizeof(kiss_twiddle_cpx)*twiddle_size;
 
     if (lenmem == NULL) {
-        st = (kiss_fftr_cfg) KISS_FFT_MALLOC (memneeded);
+        st = (kiss_fftr_cfg) KISS_FFT_MALLOC ((int)memneeded);
     } else {
         if (*lenmem >= memneeded)
             st = (kiss_fftr_cfg) mem;
